requirements.txt
evdev
pyserial
fastapi
python-dotenv
SQLAlchemy
setuptools